Several major changes and additions:
1) Individual firing of barrels with all fire option
2) Discovery that switch of Logitech Gamepad F310 to DirectInput allows
for use of triggers as digital inputs rather than analog
3)	Start of implementation for accumulation of air (WIP)

Some know issues involve slight incompatibility with the most up to date
JRE/JVM (1.8 at the time of development). Must go into the properties of
the project and set the Runtime Environment to Java 8 and then
Clean/Rebuild the project. If neglected, Eclipse will not recognize
"enum" as a keyword of Java (essential to the mapping of solenoids and
commands).