Everything Works!
- robot can shoot all at once and shoot each barrel at a time
- robot can drive and shoot at the same time
Basically the way the code works is that instead of a particular fire
command opening a solenoid then closing it, there are two separate
commands--one for opening and one for closing. the open command is run
when the button for the barrel is active and the close command is run
when the button for the barrel is inactive. has already been tested.